% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_lag.r
\name{apply_lag}
\alias{apply_lag}
\title{apply_lag}
\usage{
apply_lag(data, hold, move, lag = 0)
}
\arguments{
\item{data}{a \code{\link[base]{data.frame}} to operate over}

\item{hold}{a vector of column names to hold steady}

\item{move}{a vector of column names to lag}

\item{lag}{the number of rows to lag by}
}
\value{

}
\description{
lag some of the columns in a \code{\link[base]{data.frame}}
}
\examples{
  \dontshow{
    library(mndredge)}
  data.frame(
    a = c(1,2,3,4),
    b = c(2,3,4,5),
    c = c(4,5,6,7)) \%>\%
    apply_lag(hold = "a", move = "b", lag = 1)

}
\seealso{
Other data_dredging: \code{\link{data_dredge_correlation}},
  \code{\link{data_dredge_regression}}
}
\author{
Mark Newman, \email{mark@trinetteandmark.com}
}
\keyword{macros}
