% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dredge_correlation.r
\name{data_dredge_correlation}
\alias{data_dredge_correlation}
\title{data_dredge_correlation}
\usage{
data_dredge_correlation(data, max_lag = 0, max_pvalue = 0.05, ...)
}
\arguments{
\item{data}{a \code{\link[base]{data.frame}} to operate over}

\item{max_lag}{the maximum lag to search}

\item{max_pvalue}{the maximum the pvalue from \code{\link[stats]{cor.test}} can be before the result is filtered out}

\item{...}{addtional parameters to pass to \code{\link[stats]{cor.test}}}
}
\value{

}
\description{
data dredge a \code{\link[base]{data.frame}} using \code{\link[stats]{cor.test}}
}
\examples{
  \dontshow{
    library(mndredge)}
  set.seed(0)
  y1 <- seq(-10, 10, .5) \%>\% as.integer()
  y2 <- y1 + y1 \%>\% length() \%>\% rnorm()
  data.frame(y1, y2) \%>\% data_dredge_correlation()

}
\seealso{
Other data_dredging: \code{\link{apply_lag}},
  \code{\link{data_dredge_regression}}
}
\author{
Mark Newman, \email{mark@trinetteandmark.com}
}
\keyword{macros}
