% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dredge_regression.r
\name{data_dredge_regression}
\alias{data_dredge_regression}
\title{data_dredge_regression}
\usage{
data_dredge_regression(data, max_columns = 4, min_pvalue = 1e-10,
  max_pvalue = 0.05, ...)
}
\arguments{
\item{data}{a \code{\link[base]{data.frame}} to operate over}

\item{max_columns}{the maximum columns that can go into \code{\link[stats]{lm}} formula}

\item{min_pvalue}{the maximum the pvalue from \code{\link[stats]{lm}} can be before the result is filtered out. this generaly is an issue only in overfitted models}

\item{max_pvalue}{the maximum the pvalue from \code{\link[stats]{lm}} can be before the result is filtered out}

\item{...}{addtional parameters to pass to \code{\link[stats]{lm}}}
}
\value{

}
\description{
data dredge a \code{\link[base]{data.frame}} using \code{\link[stats]{lm}}
}
\examples{
  \dontshow{
    library(mndredge)}
  set.seed(0)

}
\seealso{
Other data_dredging: \code{\link{apply_lag}},
  \code{\link{data_dredge_correlation}}
}
\author{
Mark Newman, \email{mark@trinetteandmark.com}
}
\keyword{macros}
